Extract the theme switcher into a visible Appearance card on Settings → General
The theme picker (System / Light / Dark, plus Velvet behind its flag) was only
reachable by opening the Preferences → Customize modal. Pull it out into its own
always-visible AppearanceCard rendered directly on the General settings page so
the theme control is no longer hidden in a modal.

Because the theme picker was the only section that gave the Preferences modal
content on touch/non-Electron surfaces (the rest is Electron-only shortcuts and
Launch at Login, or the fine-pointer-only composer toggle), gate the Preferences
card behind `isElectron() || !isPointerCoarse()` so it can't open an empty modal.

Move the theme coverage into appearance-card.test.tsx and repurpose
preferences-modal.test.tsx into a guard that the modal no longer hosts Appearance
while still rendering its remaining content.
